Split LLVM function and op emission
